% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perusirtod-package.R
\docType{package}
\name{perusirtod-package}
\alias{perusirtod}
\alias{perusirtod-package}
\title{perusirtod: An Interface to the peruvian 'SIRTOD' data}
\description{
Functionality to read data from the peruvian 'SIRTOD' (Regional Information System for Decision Making).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/calderonsamuel/perusirtod}
  \item \url{https://calderonsamuel.github.io/perusirtod/}
  \item Report bugs at \url{https://github.com/calderonsamuel/perusirtod/issues}
}

}
\author{
\strong{Maintainer}: Samuel Calderon \email{samuel.calderon@uarm.pe} (\href{https://orcid.org/0000-0001-6847-1210}{ORCID}) [copyright holder]

}
\keyword{internal}
