% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirtod_lugares.R
\docType{data}
\name{sirtod_lugares}
\alias{sirtod_lugares}
\title{Places metadata}
\format{
\subsection{\code{sirtod_lugares}}{

A data frame with 1,877 rows and 4 columns:
\describe{
\item{ubigeo}{Geographic location code. Key against \code{sirtod_valores}}
\item{departamento}{Department name}
\item{provincia}{Province name}
\item{distrito}{District name}
}
}
}
\source{
\url{https://systems.inei.gob.pe/SIRTOD/}
}
\usage{
sirtod_lugares
}
\description{
Metadata of the geographic location of the SIRTOD measurement.
It is meant to be joined/merged with \code{sirtod_valores}
}
\keyword{datasets}
