% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirtod_valores.R
\docType{data}
\name{sirtod_valores}
\alias{sirtod_valores}
\title{Measurement values}
\format{
\subsection{\code{sirtod_valores}}{

A data frame with 15,742,020 rows and 4 columns:
\describe{
\item{codigoIndicador}{Variable code. Key against \code{sirtod_indicadores}}
\item{ubigeo}{Geographic location code. Key against \code{sirtod_valores}}
\item{año}{Year of the measurement}
\item{dato}{Value of the measurement}
}
}
}
\source{
\url{https://systems.inei.gob.pe/SIRTOD/}
}
\usage{
sirtod_valores
}
\description{
Measurement values available in the SIRTOD data collection.
They are meant to be joined/merged with \code{sirtod_indicadores} or
\code{sirtod_lugares} after choosing the variables or places of interest.
}
\keyword{datasets}
